More network stubs, fix gameobject serialization
GameObjects should now serialize to the correct byte layout.  Also,
ObjectId was changed to Ushort, as there should never be 65535+ of
them...

+++ MPTanks-MK5/MPTanks.Engine/GameCore.cs
@@ -97,11 +97,11 @@ namespace MPTanks.Engine
         /// The physics world that the game runs in.
         /// </summary>
         internal FarseerPhysics.Dynamics.World World { get; private set; }
-        private int _nextObjectId = 0;
+        private ushort _nextObjectId = 0;
         /// <summary>
         /// The next available object id to use
         /// </summary>
-        internal int NextObjectId { get { return _nextObjectId++; } }
+        internal ushort NextObjectId { get { return _nextObjectId++; } }
 
         private Dictionary<int, GameObject> _gameObjects = new Dictionary<int, GameObject>();
         /// <summary>
